+++ cpp/pybind/visualization/o3dvisualizer.cpp
@@ -267,9 +267,9 @@ void pybind_o3dvisualizer_definitions(py::module& m) {
                  "Removes the geometry with the name.", "name"_a)
             .def("update_geometry", &O3DVisualizer::UpdateGeometry,
                  "Updates the attributes of the named geometry specified by "
-                 "update_flags with tpoint_cloud. Note: Currently this "
-                 "function only works with T Geometry Point Clouds.",
-                 "name"_a, "tpoint_cloud"_a, "update_flags"_a)
+                 "update_flags from the tensor PointCloud or fixed-topology "
+                 "TriangleMesh.",
+                 "name"_a, "tgeometry"_a, "update_flags"_a)
             .def("show_geometry", &O3DVisualizer::ShowGeometry,
                  "Checks or unchecks the named geometry in the list. Note that "
                  "even if show_geometry(name, True) is called, the object may "